requirements.txt
appnope==0.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-cors-headers==3.2.0
django-crispy-forms==1.8.0
django-filter==2.2.0
django-formtools==2.1
django-guardian==2.1.0
django-haystack==2.8.1
django-import-export==1.2.0
django-qiniu-storage==2.3.1
django-redis==4.10.0
django-reversion==3.0.4
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-extensions==0.5.0
drf-haystack==1.8.6
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.0
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
itypes==1.1.0
jdcal==1.4.1
jedi==0.15.1
Jinja2==2.10.3
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.1.1
mysqlclient==1.4.6
oauthlib==3.1.0
odfpy==1.4.0
olefile==0.46
openpyxl==3.0.1
parso==0.5.1
pep562==1.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
prompt-toolkit==2.0.10
ptyprocess==0.6.0
pycryptodome==3.9.4
Pygments==2.4.2
PyJWT==1.7.1
pymdown-extensions==6.2
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.1.2
qiniu==7.2.6
raven==6.10.0
redis==3.3.11
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
tablib==0.14.0
traitlets==4.3.3
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
webencodings==0.5.1
Whoosh==2.7.4
xlrd==1.2.0
XlsxWriter==1.2.6
xlwt==1.3.0
